+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Fix crash in `FFmpegFrameGrabber.restart()` ([issue #605](https://github.com/bytedeco/javacv/issues/605))
  * Upgrade dependencies for OpenCV 3.2.0, libdc1394 2.2.5
 
 ### January 14, 2017 version 1.3.1